requirements.txt
yt-dlp>=2023.11.16
Pillow>=10.0.0
requests>=2.31.0 